chore: optimize

+++ packages/client-api/src/requests/transfer.ts
@@ -3,6 +3,7 @@ import { Transfer, IResponse } from '@lib/types';
 import { error, transfer as types } from '../types';
 import { BaseApi } from '../types/BaseApi';
 import { BaseRequest } from '../types/BaseRequest';
+import { sleep } from '../utils';
 
 class CTransfer extends BaseRequest {
   constructor(api: BaseApi) {
@@ -10,6 +11,15 @@ class CTransfer extends BaseRequest {
   }
 
   setTransfer = async () => {
+    if (this.api.config.debug?.isMock) {
+      await sleep(this.api.config.debug?.mockDelay || 0);
+
+      return {
+        type: 'SET_TRANSFER',
+        status: undefined,
+      } as types.ISetTransferResponse;
+    }
+
     try {
       const res = await this.api.axios.post<IResponse<Transfer>>('/transfer');
       const resData = res.data;
@@ -34,6 +44,15 @@ class CTransfer extends BaseRequest {
   };
 
   getTransfer = async () => {
+    if (this.api.config.debug?.isMock) {
+      await sleep(this.api.config.debug?.mockDelay || 0);
+
+      return {
+        type: 'GET_TRANSFER',
+        status: undefined,
+      } as types.IGetTransferResponse;
+    }
+
     try {
       const res = await this.api.axios.get<IResponse<Transfer>>('/transfer');
       const resData = res.data;
@@ -57,15 +76,23 @@ class CTransfer extends BaseRequest {
     }
   };
 
-  removeMessage = async (uid: string) => {
+  clearTransfer = async (uid: string) => {
+    if (this.api.config.debug?.isMock) {
+      await sleep(this.api.config.debug?.mockDelay || 0);
+
+      return {
+        type: 'CLEAR_TRANSFER',
+      } as types.IClearTransferResponse;
+    }
+
     try {
       const res = await this.api.axios.delete<IResponse<void>>(`/transfer/${uid}`);
       const resData = res.data;
 
       if (resData.result) {
         return {
-          type: 'REMOVE_TRANSFER',
-        } as types.IRemoveTransferResponse;
+          type: 'CLEAR_TRANSFER',
+        } as types.IClearTransferResponse;
       }
       return {
         type: 'ERROR',
